Moved mixins inside packages, made players immune to suffocation damage while they have ForestAuraEffect
