+++ src/main/java/me/emafire003/dev/lightwithin/mixin/light_triggers/EntityDeathMixin.java
@@ -1,4 +1,4 @@
-package me.emafire003.dev.lightwithin.mixin;
+package me.emafire003.dev.lightwithin.mixin.light_triggers;
 
 import me.emafire003.dev.lightwithin.events.EntityDeathEvent;
 import net.minecraft.entity.LivingEntity;
