+++ src/main/java/me/emafire003/dev/lightwithin/mixin/light_triggers/LivingEntityDamageTriggersMixin.java
@@ -1,4 +1,4 @@
-package me.emafire003.dev.lightwithin.mixin;
+package me.emafire003.dev.lightwithin.mixin.light_triggers;
 
 import me.emafire003.dev.lightwithin.events.EntityAttackEntityEvent;
 import me.emafire003.dev.lightwithin.events.EntityBurningEvent;
