+++ src/main/resources/lightwithin.mixins.json
@@ -5,29 +5,30 @@
   "compatibilityLevel": "JAVA_16",
   "mixins": [
     "AnvilItemCrushMixin",
-    "ArmorOrItemBreakMixin",
+    "light_triggers.ArmorOrItemBreakMixin",
     "BottleLightMixin",
     "DrownedEntityMixin",
-    "EnterForestBlocksEntityMixin",
-    "EntityAttackEntityMixin",
-    "EntityDeathMixin",
-    "EntityFallingMixin",
-    "ForestBlocksMixin",
+    "forest_aura_related.EnterForestBlocksEntityMixin",
+    "light_triggers.EntityAttackEntityMixin",
+    "light_triggers.EntityDeathMixin",
+    "light_triggers.EntityFallingMixin",
+    "forest_aura_related.ForestBlocksMixin",
     "FreezeEntityLookMixin",
     "FreezeEntityMixin",
-    "LivingEntityDamageTriggersMixin",
+    "light_triggers.LivingEntityDamageTriggersMixin",
     "MilkNotCleanLightEffectsMixin",
     "PlayerFreezeMixin",
     "PlayerJoinMixin",
-    "PreventLightEffectClearLivingEntityMixin"
+    "PreventLightEffectClearLivingEntityMixin",
+    "forest_aura_related.SuffocationImmunityEntityMixin"
   ],
   "client": [
-    "BackgroundRenderForestMixin",
-    "CameraMixin",
+    "forest_aura_related.BackgroundRenderForestMixin",
+    "forest_aura_related.CameraMixin",
     "CustomEffectsLivingRendererMixin",
-    "EntitiesGlowingForClientMixin",
-    "GameRendererMixin",
-    "InGameHudMixin",
+    "forest_aura_related.EntitiesGlowingForClientMixin",
+    "renderer.GameRendererMixin",
+    "renderer.InGameHudMixin",
     "PlayerChargedGlowRendererMixin",
     "RenderEffectsEntityMixin"
   ],
